refactor: split web create event form
