+++ apps/web/src/features/create-event/types.ts
@@ -1,0 +1,12 @@
+export type SelectedEventLocation = {
+  kind: 'current' | 'place';
+  label: string;
+  address?: string;
+  latitude: number;
+  longitude: number;
+};
+
+export type EventTagOption = {
+  id: string;
+  name: string;
+};
