+++ apps/web/src/features/create-event/utils/location.ts
@@ -1,0 +1,6 @@
+const LOCATION_COORDINATE_DECIMALS = 7;
+
+export function roundGeographicCoordinate(value: number): number {
+  const factor = 10 ** LOCATION_COORDINATE_DECIMALS;
+  return Math.round(value * factor) / factor;
+}
